+++ react/src/index.ts
@@ -12,5 +12,8 @@ export {
 } from "./hooks/useRecords";
 export { useReverseLookup } from "./hooks/useReverseLookup";
 export { useSnsDomainsForOwner } from "./hooks/useSnsDomainsForOwner";
+export { useSnsNftsForOwner } from "./hooks/useSnsNftsForOwner";
+export { useSolDomainsForOwner } from "./hooks/useSolDomainsForOwner";
+export { useSolNftsForOwner } from "./hooks/useSolNftsForOwner";
 export { useSubdomains } from "./hooks/useSubdomains";
 export type { Options } from "./types";